# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ShopFlow Platform API
  description: |
    REST API for the wholesale/retail shop management platform (React web and React Native clients).

    * Authentication: mobile number + OTP → short-lived bearer access token (JWT) plus a rotating refresh token.
    * Authorization: role (OWNER, ADMIN, CUSTOMER) plus permission codes; customers can only access their own data.
    * Envelope: every JSON response is `{ success, data, message, pagination?, error?, requestId }`.
    * Errors: `error.code` is one of the documented business codes (e.g. INSUFFICIENT_STOCK, CREDIT_LIMIT_EXCEEDED).
    * Money: decimal numbers with 2 places, calculated only by the backend.
    * Lists: `page` (1-based), `pageSize` (max 100), `sort=field,asc|desc`.
    * Retries: send `Idempotency-Key` on order creation, payment recording, invoice generation and WhatsApp sends.
  version: v1
servers:
- url: /
  description: Current host
...
paths:
...
  /api/v1/invoices:
    get:
      tags:
      - Invoices
      summary: List invoices
      description: "Customers see only their own generated invoices. Filter by status,\
        \ source, customer, overdue and date range. Sort: invoiceDate, invoiceNumber,\
        \ grandTotal."
      operationId: list_7
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: customerId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - DRAFT
          - GENERATED
          - SENT
          - PARTIALLY_PAID
          - PAID
          - CREDIT
          - CANCELLED
      - name: source
        in: query
        required: false
        schema:
          type: string
          enum:
          - ORDER
          - MANUAL
      - name: overdue
        in: query
        required: false
        schema:
          type: boolean
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseListInvoiceResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Invoices
      summary: Create an invoice (draft)
      description: "With orderId: bills the order's uninvoiced accepted (or delivered,\
        \ per settings) quantities. Without: an admin-created invoice for customerId\
        \ with items and paymentType. Set generate=true to generate immediately. Errors:\
        \ ORDER_INVALID_STATUS, NOTHING_TO_INVOICE, PRODUCT_INACTIVE, CUSTOMER_NOT_APPROVED."
      operationId: create_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseInvoiceResponse"
      security:
      - bearerAuth: []
  /api/v1/invoices/{id}/send-whatsapp:
    post:
      tags:
      - Invoices
      summary: Send the invoice on WhatsApp
      description: Queues delivery of the PDF to the customer's mobile. Delivery is
        asynchronous and retried; status is tracked per message. Send Idempotency-Key
        to avoid duplicate sends.
      operationId: sendWhatsApp
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseWhatsAppMessageResponse"
      security:
      - bearerAuth: []
  /api/v1/invoices/{id}/generate:
    post:
      tags:
      - Invoices
      summary: Generate (post) a draft invoice
      description: "Assigns the invoice number, snapshots seller/buyer, posts the\
        \ customer ledger, applies advances and (for admin-created invoices) posts\
        \ stock OUT — atomically. Idempotent; send Idempotency-Key for safe retries.\
        \ Errors: CREDIT_LIMIT_EXCEEDED, INSUFFICIENT_STOCK."
      operationId: generate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseInvoiceResponse"
      security:
      - bearerAuth: []
  /api/v1/invoices/{id}/credit-notes:
    post:
      tags:
      - Invoices
      summary: Issue a credit note
      description: For price adjustments or other corrections. Sales returns create
        credit notes automatically on approval.
      operationId: creditNote
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditNoteRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseInvoiceResponse"
      security:
      - bearerAuth: []
  /api/v1/invoices/{id}/cancel:
    post:
      tags:
      - Invoices
      summary: Cancel an invoice
      description: "Never deletes. Reverses the ledger, turns allocated payments into\
        \ customer advances and returns stock for admin-created invoices. Errors:\
        \ INVOICE_ALREADY_CANCELLED."
      operationId: cancel_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelInvoiceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseInvoiceResponse"
      security:
      - bearerAuth: []
  /api/v1/invoices/whatsapp-messages/{messageId}/retry:
    post:
      tags:
      - Invoices
      summary: Retry a failed WhatsApp message
      operationId: retryWhatsApp
      parameters:
      - name: messageId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseWhatsAppMessageResponse"
      security:
      - bearerAuth: []
...
  /api/v1/business/settings:
    get:
      tags:
      - Business Settings
      summary: Get operational business rules
      description: "Credit policy, cancellation window, partial-delivery invoicing,\
        \ notifications, retention."
      operationId: settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBusinessSettingsResponse"
      security:
      - bearerAuth: []
    patch:
      tags:
      - Business Settings
      summary: Update operational business rules
      operationId: updateSettings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessSettingsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBusinessSettingsResponse"
      security:
      - bearerAuth: []
...
  /api/v1/orders/{id}/status-history:
    get:
      tags:
      - Orders
      summary: Order status timeline
      operationId: history
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseListStatusHistoryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/my/outstanding:
    get:
      tags:
      - My Account
      summary: My outstanding and credit balance
      operationId: outstanding
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOutstanding"
      security:
      - bearerAuth: []
...
  /api/v1/invoices/{id}:
    get:
      tags:
      - Invoices
      summary: "Get an invoice with lines, tax summary and credit notes"
      operationId: get_10
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseInvoiceResponse"
      security:
      - bearerAuth: []
  /api/v1/invoices/{id}/whatsapp-messages:
    get:
      tags:
      - Invoices
      summary: WhatsApp delivery history for an invoice
      operationId: whatsAppMessages
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseListWhatsAppMessageResponse"
      security:
      - bearerAuth: []
  /api/v1/invoices/{id}/pdf:
    get:
      tags:
      - Invoices
      summary: Download the invoice PDF
      description: "A4, print-ready. Generated invoices return the stored, immutable\
        \ PDF."
      operationId: pdf
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/pdf:
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
...
  /api/v1/dashboard/customer:
    get:
      tags:
      - Dashboard
      summary: Customer home summary
      description: "Outstanding, open orders, recent orders and recently ordered products\
        \ for the signed-in customer."
      operationId: customer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
      security:
      - bearerAuth: []
...
  /api/v1/customers/{id}/outstanding:
    get:
      tags:
      - Customers
      summary: Outstanding and credit summary
      operationId: outstanding_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseOutstanding"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldIssue"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    CreateInvoiceRequest:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        paymentType:
          type: string
          enum:
          - CASH
          - ONLINE
          - UPI
          - BANK_TRANSFER
          - CREDIT
          - OTHER
        invoiceDate:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineRequest"
          maxItems: 200
          minItems: 0
        paymentTerms:
          type: string
          maxLength: 200
          minLength: 0
        buyerOrderNumber:
          type: string
          maxLength: 60
          minLength: 0
        deliveryNote:
          type: string
          maxLength: 200
          minLength: 0
        dispatchDocument:
          type: string
          maxLength: 200
          minLength: 0
        transport:
          type: string
          maxLength: 200
          minLength: 0
        vehicleNumber:
          type: string
          maxLength: 20
          minLength: 0
        destination:
          type: string
          maxLength: 200
          minLength: 0
        notes:
          type: string
          maxLength: 1000
          minLength: 0
        generate:
          type: boolean
    InvoiceLineRequest:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: number
          minimum: 0.001
        rate:
          type: number
          minimum: 0
        discountPercent:
          type: number
          maximum: 100
          minimum: 0
        discountAmount:
          type: number
          minimum: 0
        description:
          type: string
          maxLength: 500
          minLength: 0
      required:
      - productId
      - quantity
    ApiResponseInvoiceResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/InvoiceResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
...
    ApiResponseWhatsAppMessageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/WhatsAppMessageResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    WhatsAppMessageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoiceId:
          type: string
          format: uuid
        recipient:
          type: string
        status:
          type: string
        providerMessageId:
          type: string
        failureReason:
          type: string
        retryCount:
          type: integer
          format: int32
        nextRetryAt:
          type: string
          format: date-time
        queuedAt:
          type: string
          format: date-time
        sentAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
    CreateCreditNoteRequest:
      type: object
      properties:
        reasonType:
          type: string
          enum:
          - SALES_RETURN
          - SHORT_DELIVERY
          - DELIVERY_FAILED
          - PRICE_ADJUSTMENT
          - OTHER
        reason:
          type: string
          maxLength: 500
          minLength: 0
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreditNoteLineRequest"
          maxItems: 2147483647
          minItems: 1
      required:
      - items
      - reason
      - reasonType
    CreditNoteLineRequest:
      type: object
      properties:
        invoiceItemId:
          type: string
          format: uuid
        quantity:
          type: number
          minimum: 0.001
      required:
      - invoiceItemId
      - quantity
    CancelInvoiceRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
          minLength: 0
      required:
      - reason
...
    UpdateBusinessSettingsRequest:
      type: object
      properties:
        creditPolicy:
          type: string
          enum:
          - BLOCK
          - REQUIRE_ADMIN_APPROVAL
          - ALLOW
        customerCancelAllowedUntil:
          type: string
          enum:
          - PLACED
          - ACCEPTED
          - NEVER
        showStockToCustomers:
          type: boolean
        gstinRequiredForCustomers:
          type: boolean
        panRequiredForCustomers:
          type: boolean
        partialDeliveryInvoicePolicy:
          type: string
          enum:
          - INVOICE_ACCEPTED_QUANTITY
          - INVOICE_DELIVERED_QUANTITY
        defaultCreditDays:
          type: integer
          format: int32
          maximum: 365
          minimum: 0
        defaultCreditLimit:
          type: number
          minimum: 0
        whatsappEnabled:
          type: boolean
        whatsappSender:
          type: string
          maxLength: 40
          minLength: 0
        whatsappInvoiceTemplate:
          type: string
          maxLength: 100
          minLength: 0
        notifyPush:
          type: boolean
        notifySms:
          type: boolean
        notifyEmail:
          type: boolean
        notifyWhatsapp:
          type: boolean
        dataRetentionYears:
          type: integer
          format: int32
          maximum: 50
          minimum: 1
    ApiResponseBusinessSettingsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/BusinessSettingsResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
...
    ApiResponseListStatusHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistoryResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
...
    ApiResponseOutstanding:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Outstanding"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
...
    ApiResponseListInvoiceResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    ApiResponseListWhatsAppMessageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/WhatsAppMessageResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    ApiResponseMapStringObject:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: {}
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT